# backend/src/http/api/v1/v1-swagger.yml
...
swagger: "2.0"
info:
  description: Documentation de l'API publique d'Anotéa
  version: 1.0.0
  title: Anotea
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
  - http
securityDefinitions:
  Authorization:
    type: apiKey
    in: header
    name: Authorization
    description: |
      Pour plus d'informations voir :
      https://github.com/StartupsPoleEmploi/anotea/blob/master/API.md#Authentification
responses:
  AvisResponse:
    description: La liste des avis
    schema:
      type: object
      properties:
        avis:
          type: array
          items:
            $ref: "#/definitions/Avis"
        meta:
          type: object
          description: Informations complémentaires
          properties:
            pagination:
              $ref: "#/definitions/Pagination"
  AvisByIdResponse:
    description: Un avis
    schema:
      $ref: "#/definitions/Avis"
...
  UnauthorizedErrorResponse:
    description: Le header Authorization header est manquant ou invalide
    schema:
      $ref: "#/definitions/Error"
  BadRequestErrorResponse:
    description: Bad Request
    schema:
      $ref: "#/definitions/Error"
  NotFoundErrorResponse:
    description: Not Found
    schema:
      $ref: "#/definitions/Error"
definitions:
  Error:
    type: object
    properties:
      statusCode:
        type: integer
        example: 5xx/4xx
      error:
        type: string
        example: Internal Server Error
      message:
        type: string
        example: An internal server error occurred
...
  Pagination:
    type: object
    description: Statut de la pagination
    properties:
      page:
        type: number
      items_par_page:
        type: number
      total_items:
        type: number
      total_pages:
        type: number
...
  Notes:
    type: object
    description: Moyenne des notes pour les avis
    properties:
      accueil:
        type: number
      contenu_formation:
        type: number
      equipe_formateurs:
        type: number
      moyen_materiel:
        type: number
      accompagnement:
        type: number
      global:
        type: number
...
  Avis:
    type: object
    properties:
      id:
        type: string
        description: L'identifiant unique de l'avis
      pseudo:
        type: string
        description: Le pseudo du stagiaire
      date:
        type: string
        description: La date à laquelle l'avis a été crée
      commentaire:
        type: object
        properties:
          titre:
            type: string
          texte:
            type: string
      reponse:
        type: object
        description: La réponse de l'organisme formateur
        properties:
          texte:
            type: string
      notes:
        $ref: "#/definitions/Notes"
      formation:
        description: La formation suivie par le stagiaire
        type: object
        properties:
          intitule:
            type: string
          domaine_formation:
            type: object
            properties:
              formacodes:
                type: array
                items:
                  type: string
          certifications:
            type: array
            uniqueItems: true
            minItems: 1
            items:
              properties:
                certif_info:
                  type: string
          action:
            type: object
            properties:
              lieu_de_formation:
                type: object
                properties:
                  code_postal:
                    type: string
                  ville:
                    type: string
              organisme_financeurs:
                type: array
                items:
                  type: string
              organisme_formateur:
                type: object
                properties:
                  raison_sociale:
                    type: string
                  siret:
                    type: string
                  numero:
                    type: string
              session:
                type: object
                properties:
                  numero:
                    type: string
                  periode:
                    type: object
                    properties:
                      debut:
                        type: string
                      fin:
                        type: string
...
paths:
  /api/v1/avis:
    get:
      description: Renvoie tous les avis triés par date et filtrés en fonction des paramètres
      operationId: getAvis
      produces:
        - application/json
      parameters:
        - in: query
          name: organisme_formateur
          description: Le numéro SIRET de l'organisme formateur
          type: string
          required: false
        - in: query
          name: formacode
          description: Le formacode de la formation (Français appliqué=15226) ou un domaine de formation (Langues=152)
          type: string
          required: false
        - in: query
          name: certif_info
          description: Le certif info de la formation
          type: string
          required: false
        - in: query
          name: lieu_de_formation
          description: Le code postal du lieu de formation
          type: string
          required: false
        - in: query
          name: page
          description: Le numéro de la page
          default: 0
          minimum: 0
          type: number
          required: false
        - in: query
          name: items_par_page
          description: Le nombre d'avis par page
          default: 50
          maximum: 2000
          type: number
          required: false
        - in: query
          name: notes_decimales
          default: false
          description: Si true, renvoie les notes en valeur décimale arrondie à la dizaine
          type: number
          required: false
      responses:
        200:
          $ref: "#/responses/AvisResponse"
        400:
          $ref: "#/responses/BadRequestErrorResponse"
        401:
          $ref: "#/responses/UnauthorizedErrorResponse"
      tags:
        - Publique
  /api/v1/avis/:id:
    get:
      description: >
        Permet d'obtenir un avis à partir de son identifiant
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: notes_decimales
          default: false
          description: Si true, renvoie les notes en valeur décimale arrondie à la dizaine
          type: number
          required: false
      security:
        - Authorization: []
      responses:
        200:
          $ref: "#/responses/AvisByIdResponse"
        400:
          $ref: "#/responses/BadRequestErrorResponse"
        401:
          $ref: "#/responses/UnauthorizedErrorResponse"
        404:
          $ref: "#/responses/NotFoundErrorResponse"
      tags:
        - Publique
...